// src/api/generated/model/orderDto.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * ECommerce API
 * ECommerce API with OpenIddict Authentication and Versioning
 * OpenAPI spec version: v1
 */
import type { OrderStatus } from './orderStatus';
import type { OrderItemDto } from './orderItemDto';

export interface OrderDto {
  id?: string;
  userId?: string;
  orderDate?: string;
  status?: OrderStatus;
  totalAmount?: number;
  /** @nullable */
  shippingAddress?: string | null;
  /** @nullable */
  billingAddress?: string | null;
  /** @nullable */
  items?: OrderItemDto[] | null;
}