// src/api/generated/model/orderStatusUpdateCommand.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * ECommerce API
 * ECommerce API with OpenIddict Authentication and Versioning
 * OpenAPI spec version: v1
 */
import type { OrderStatus } from './orderStatus';

export interface OrderStatusUpdateCommand {
  orderId?: string;
  newStatus?: OrderStatus;
}